% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlf_IdahoFeas_data.R
\docType{data}
\name{dat_alternatives_to_run}
\alias{dat_alternatives_to_run}
\title{Management alternatives that re to be evaluated as part of this feasibility assessment.}
\format{
An object of class \code{data.frame} with 12 rows and 11 columns.
}
\source{
Group framing facilitation, April 1 2021
}
\usage{
dat_alternatives_to_run
}
\description{
A table containing the alternatives that can be run with this model. 
These alternatives were generated during facilitated group framing sessions, 
with the last one where we finalized these alternatives occurring on April 2, 2021.
}
\keyword{datasets}
