% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlf_IdahoFeas_dispersal.R
\name{dispersalTracking}
\alias{dispersalTracking}
\title{Applies the dispersal models and returns a matrix showing how many yoy dispersed 
from one wetland cell to another.}
\usage{
dispersalTracking(
  resultsTracking_popSize_females,
  yoy_rows,
  i,
  j,
  wetlands,
  wetlands_not_dry,
  wetland_distances_km,
  parameterByIterTracking,
  allow_outside,
  demographic_stochasticity = TRUE
)
}
\arguments{
\item{resultsTracking_popSize_females}{Formatted as per the function dapva::makeResultsTracking()}

\item{yoy_rows}{which(resultsTracking_popSize_females$class == "yoy")}

\item{i}{Iteration #}

\item{j}{Year}

\item{wetlands}{A vector of the wetland names that are available in the WMA.}

\item{wetlands_not_dry}{A vector of the wetland names that are available as dipersal destinations this year becasue they are not dry.}

\item{wetland_distances_km}{A matrix of distance between the wetlands.}

\item{parameterByIterTracking}{A dataframe of the form parameterByIterTracking()
that contains the following dispersal related inputs: dispersal_CSF_vs_MoreGoShort,
dispersal_CSFmodel_lessEqual1km, etc.}

\item{allow_outside}{"yes" or "no", allow frogs to disperse outside of the WMA or not.}

\item{demographic_stochasticity}{TRUE or FALSE, defaults to TRUE}
}
\description{
The rows are the 'from' wetland and the columns are the 'to' wetland/location.
}
\examples{
resultsTracking_popSize_females <- dapva::makeResultsTracking(i = 1,
                                                             yrs = 50, 
                                                             initial_year = 1,
                                                             pops = c("cell3", "cell4", "cell7"), 
                                                             class_names = c("eggs", "tadpoles", 
                                                                             "yoy", "juv", "A2", 
                                                                             "A3", "A4plus"),
                                                             sex = "female")

resultsTracking_popSize_females[,paste("1")] <- 100 # put 100 in each wetland and age class

yoy_rows <- which(resultsTracking_popSize_females$class == "yoy")

inputs_all <- dapva4nlf::getNLFIdahoFeasinputs()
inputs <- inputs_all[[1]]
wetland_distances_km <- inputs_all[[2]]

parameterByIterTracking_baseCase <- dapva4nlf::selectNLFIdahoParameterByIterTracking(inputs, base_case = TRUE)

# Run the dispersal function
dispersal_results <- dispersalTracking(resultsTracking_popSize_females, 
                                      yoy_rows, i = 1, j = 1,
                                      wetlands = c("cell3", "cell4", "cell7"),
                                      wetlands_not_dry = c("cell3", "cell4", "cell7"),
                                      wetland_distances_km,
                                      parameterByIterTracking = parameterByIterTracking_baseCase,
                                      allow_outside = 'no', 
                                      demographic_stochasticity = FALSE)

}
