% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlf_IdahoFeas_inputs.R
\name{getNLFIdahoFeasinputs}
\alias{getNLFIdahoFeasinputs}
\title{Get inputs for the NLF Idaho Feasability assessment.}
\usage{
getNLFIdahoFeasinputs()
}
\value{
Returns a list of all the inputs needed. See example.
}
\description{
This is an all-encompassing function that gets the inputs needed
to run the NLF Idaho Feasability Assessment.
}
\examples{
test <- getNLFIdahoFeasinputs()
# no unit test since very specific; appears to be working correctly
}
