% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlf_IdahoFeas_inputs.R
\name{makeTornadoParameterLabels}
\alias{makeTornadoParameterLabels}
\title{Make a dataframe with clean labels for the tornado.}
\usage{
makeTornadoParameterLabels(parameterByIterTracking)
}
\arguments{
\item{parameterByIterTracking}{Output from selectNLFIdahoParameterByIterTracking()}
}
\value{
Returns a dataframe with two columns - one called 'name', which matches 
the column names from parameterByIterTracking and the other called 'label'
}
\description{
Make a dataframe with clean labels for the tornado.
}
\examples{
# makeTornadoParameterLabels(parameterByIterTracking = parameterByIterTracking_this_alt_clean)

}
