% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlf_IdahoFeas_annualLoop.R
\name{runAnnualLoopNLFIdahoPVA}
\alias{runAnnualLoopNLFIdahoPVA}
\title{This functions contains the annual loop of the Idaho Feasability 
Northern Leopard Frog population model.
The inputs to this function represent one set of parameters. This function
can be called inside an iteration loop and a run loop. For each iteration, a
different set of parameters should be chosen to simulate parametric
uncertainty. For each run, the same parameters are used and differences in
the output are due to environmental and demographic stochasticity.}
\usage{
runAnnualLoopNLFIdahoPVA(
  parameterByIterTracking,
  yrs,
  i,
  q,
  initial_year,
  wetlands,
  K,
  percentilesEV
)
}
\arguments{
\item{parameterByIterTracking}{Parameters to use in each iteration in the
format of selectBTPDparameterByIterTracking().}

\item{yrs}{The number of years to run the model for (e.g. 50).}

\item{i}{The iteration number. For results tracking purposes.}

\item{q}{The run number. For results tracking purposes.}

\item{initial_year}{First year of the model.}

\item{wetlands}{Wetland cell names/abbreviations.}

\item{K}{Carrying capacity .}

\item{percentilesEV}{A vector of lenth years with percentiles to use for
environmental variation (EV) in each iteration (e.g. output from selectEVPercentilesNormal()).}
}
\description{
This functions contains the annual loop of the Idaho Feasability 
Northern Leopard Frog population model.
The inputs to this function represent one set of parameters. This function
can be called inside an iteration loop and a run loop. For each iteration, a
different set of parameters should be chosen to simulate parametric
uncertainty. For each run, the same parameters are used and differences in
the output are due to environmental and demographic stochasticity.
}
