% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlf_IdahoFeas_dispersal.R
\name{dispersalTracking}
\alias{dispersalTracking}
\title{Applies the dispersal models and returns a matrix showing how many yoy dispersed 
from one wetland cell to another.}
\usage{
dispersalTracking(
  resultsTracking_popSize_females,
  yoy_rows,
  i,
  j,
  wetlands,
  wetland_distances_km,
  parameterByIterTracking
)
}
\arguments{
\item{i}{Iteration #}

\item{j}{Year}

\item{wetlands}{A vector of the wetland names that are available in the WMA.}

\item{wetland_distances_km}{A matrix of distance between the wetlands.}

\item{parameterByIterTracking}{A dataframe of the form parameterByIterTracking()
that contains the following dispersal related inputs: dispersal_CSF_vs_MoreGoShort,
dispersal_CSFmodel_lessEqual1km, etc.}
}
\description{
The rows are the 'from' wetland and the columns are the 'to' wetland/location.
}
\details{
INSERT
}
\examples{
# Still to do

}
