% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlf_IdahoFeas_annualLoop.R
\name{runAnnualLoopNLFIdahoPVA}
\alias{runAnnualLoopNLFIdahoPVA}
\title{This functions contains the annual loop of the Idaho Feasability 
Northern Leopard Frog population model.
The inputs to this function represent one set of parameters. This function
can be called inside an iteration loop and a run loop. For each iteration, a
different set of parameters should be chosen to simulate parametric
uncertainty. For each run, the same parameters are used and differences in
the output are due to environmental and demographic stochasticity.}
\usage{
runAnnualLoopNLFIdahoPVA(
  parameterByIterTracking,
  yrs,
  i,
  q,
  wetland_distances_km,
  initial_year,
  wetlands,
  stage_classes,
  percentilesEV_survival_eggs_tad,
  percentilesEV_survival_yoy_adult,
  percentilesEV_reproduction,
  alternative_details
)
}
\arguments{
\item{parameterByIterTracking}{Parameters to use in each iteration in the
format of selectBTPDparameterByIterTracking().}

\item{yrs}{The number of years to run the model for (e.g. 50).}

\item{i}{The iteration number. For results tracking purposes.}

\item{q}{The run number. For results tracking purposes.}

\item{wetland_distances_km}{Matrix of distances between wetlands in kilometers.}

\item{initial_year}{First year of the model.}

\item{wetlands}{Wetland cell names/abbreviations.}

\item{stage_classes}{ADD DESCRIPTION LATER}

\item{percentilesEV_survival_eggs_tad}{Columns are the different wetlands, rows 
are a vector of length years with percentiles to use for environmental variation (EV) 
in each iteration (e.g. output from selectEVPercentilesNormal()).}

\item{percentilesEV_survival_yoy_adult}{Columns are the different wetlands, rows 
are a vector of length years with percentiles to use for environmental variation (EV) 
in each iteration (e.g. output from selectEVPercentilesNormal())..}

\item{percentilesEV_reproduction}{Columns are the different wetlands, rows 
are a vector of length years with percentiles to use for environmental variation (EV) 
in each iteration (e.g. output from selectEVPercentilesNormal()).}

\item{alternative_details}{ADD DESCRIPTION LATER}
}
\description{
This functions contains the annual loop of the Idaho Feasability 
Northern Leopard Frog population model.
The inputs to this function represent one set of parameters. This function
can be called inside an iteration loop and a run loop. For each iteration, a
different set of parameters should be chosen to simulate parametric
uncertainty. For each run, the same parameters are used and differences in
the output are due to environmental and demographic stochasticity.
}
