% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlf_IdahoFeas_survival.R
\name{survivalOverwinter}
\alias{survivalOverwinter}
\title{Calculate overwinter survival of terrestrial (i.e. the ones that overwinter) life stages.}
\usage{
survivalOverwinter(
  popSizeVector_fallLastYear,
  survivalMatrix,
  demographic_stochasticity = TRUE
)
}
\arguments{
\item{popSizeVector_fallLastYear}{Female population sizes formatted in the style
of dapva::makePopSizeVector() for the previous year.}

\item{survivalMatrix}{Survival matrix formatted as per dapva::makeSurvivalMatrix().}

\item{demographic_stochasticity}{TRUE or FALSE, defaults to TRUE}
}
\description{
Applies the survival matrix to the fall population size vector after masking eggs 
and tadpoles (since those are intermediate stages that are grown into young-of-year by the fall)
to calculate how many of each terrestrial age class survives over winter to the spring.
}
\examples{
library(dapva)
popSizeVector_test <- dapva::makePopSizeVector(
pop_class_names = c("pop1 eggs", "pop1 tadpoles", "pop1 adults"),
num_indiv_per_pop_class = c(1000, 100, 20), sex = "female"
)
survivalMatrix_test <- dapva::makeSurvivalMatrix(
 class_names = c("pop1 eggs", "pop1 tadpoles", "pop1 adults"),
 survival_rates = c(0.5, 0.4, 0.9)
)
survived_the_winter_test <- dapva4nlf::survivalOverwinter(popSizeVector_test, 
                                                         survivalMatrix_test,
                                                         demographic_stochasticity = FALSE)

}
