% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlf_IdahoFeas_survival.R
\name{survivalTadpoles}
\alias{survivalTadpoles}
\title{INSERT}
\usage{
survivalTadpoles(
  popSizeVector,
  survivalMatrix,
  demographic_stochasticity = TRUE
)
}
\arguments{
\item{popSizeVector}{Female population sizes formatted in the style
of makePopSizeVector() for the relevant year.}

\item{survivalMatrix}{Survival matrix formatted in INSERT.}

\item{demographic_stochasticity}{TRUE or FALSE, defaults to TRUE}
}
\description{
INSERT
}
\examples{
# Still to do

}
